+++ server/controllers/cards/delete-card.js
@@ -1,0 +1,9 @@
+module.exports = async (ctx) => {
+    await ctx.cards.remove(ctx.params.id);
+
+    ctx.body = {
+        result: 'success'
+    };
+
+    ctx.status = 200;
+};
