+++ server/controllers/cards/get-cards.js
@@ -1,0 +1,3 @@
+module.exports = async (ctx) => {
+	ctx.body = await ctx.cards.getAll()
+};
