+++ server/controllers/transactions/delete-transaction.js
@@ -1,0 +1,1 @@
+module.exports = (ctx) => ctx.status = 405;
