+++ server/controllers/transactions/get-transactions.js
@@ -1,0 +1,4 @@
+module.exports = async (ctx) => {
+	ctx.body = await ctx.transactions.getByCardId(ctx.params.id);
+	ctx.status = 200;
+};
